% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add-numbers.R
\name{add_numbers}
\alias{add_numbers}
\title{test function}
\usage{
add_numbers(x, y)
}
\arguments{
\item{x}{numeric}

\item{y}{numeric}
}
\value{
numeric
}
\description{
test function
}
\examples{
add_numbers(1, 1)
add_numbers(2, 3)
}

